+++ WebChatGPT/__init__.py
@@ -2,7 +2,7 @@ from .main import ChatGPT
 
 __all__ = ["ChatGPT"]
 
-__version__ = "0.0.2"
+__version__ = "0.1.0"
 __author__ = "Smartwa"
 __repo__ = "https://github.com/Simatwa/WebChatGPT"
 __info__ = "Reverse Engineered ChatGPT Web-version."